Add deadlock prevention to GI waypoint.
This should hopefully make sure that there is always at least one
chunk in flight or one work token requested. As long as there is one of
either of those, we will called and will have a chance to request more
work tokens or produce more chunks from cache.

If ever we have no chunks in flight AND no work tokens pending, we have
officially deadlocked.