Add segment iteration, segmented property, and ChunkBoundary calling to Segmenter.
Those that take a Segmenter by state can now easily tell the GLa is a Segmenter by calling `is('segmented')` on the type object.

The segmenter can be asked for is number of states, and can be asked for a particular state by index.

The segmenter now will call ChunkBoundary on local segments before merging them into the global state if they define that method.